@@ -13,8 +13,8 @@
         }
     ],
     "require": {
-        "php": ">=5.5.0",
-        "atk4/data": "*",
+        "php": ">=5.6.0",
+        "atk4/data": "^1.1.8",
         "facebook/webdriver": "^1.1"
     },
     "require-dev": {
@@ -22,7 +22,7 @@
         "facebook/webdriver": "^1.1",
         "phpunit/phpunit-selenium": ">=1.2",
         "fzaninotto/faker": "*",
-        "atk4/core": "~1.1.6",
+        "atk4/core": "~1.1.8",
         "semantic/ui": "*",
         "codeclimate/php-test-reporter": "*"
     },
