@@ -1,57 +1,58 @@
 <?php
 /**
- * Demonstrates how to use layouts
+ * Demonstrates how to use layouts.
  */
+require '../vendor/autoload.php';
 
-require'../vendor/autoload.php';
+$layout = new \atk4\ui\Layout\App(['defaultTemplate'=>'./templates/layout1.html']);
 
-$layout = new \atk4\ui\Layout\App(['template'=>'./templates/layout1.html']);
-
-
-class Persistence_Faker extends \atk4\data\Persistence {
+class Persistence_Faker extends \atk4\data\Persistence
+{
     public $faker = null;
 
     public $count = 50;
 
-    function __construct($opts = []) {
+    public function __construct($opts = [])
+    {
         //parent::__construct($opts);
 
-        if(!$this->faker) {
+        if (!$this->faker) {
             $this->faker = Faker\Factory::create();
         }
     }
 
-    function prepareIterator($m) {
-        foreach($this->export($m) as $row) {
+    public function prepareIterator($m)
+    {
+        foreach ($this->export($m) as $row) {
             yield $row;
         }
     }
 
-    function export($m, $fields = []) {
-        if(!$fields) {
-            foreach($m->elements as $name=>$e) {
-                if($e instanceof \atk4\data\Field) {
+    public function export($m, $fields = [])
+    {
+        if (!$fields) {
+            foreach ($m->elements as $name=>$e) {
+                if ($e instanceof \atk4\data\Field) {
                     $fields[] = $name;
                 }
             }
         }
 
         $data = [];
-        for ($i = 0; $i<$this->count; $i++) {
+        for ($i = 0; $i < $this->count; $i++) {
             $row = [];
-            foreach($fields as $field) {
+            foreach ($fields as $field) {
                 $type = $field;
 
-                if($field == $m->id_field) {
-                    $row[$field] = $i+1;
+                if ($field == $m->id_field) {
+                    $row[$field] = $i + 1;
                     continue;
                 }
 
                 $actual = $m->getElement($field)->actual;
-                if($actual){ 
+                if ($actual) {
                     $type = $actual;
                 }
-
 
                 $row[$field] = $this->faker->$type;
             }
@@ -75,13 +76,10 @@ try {
     $layout->add(new \atk4\ui\Lister(), 'Report')
         ->setModel($m);
 
-
     echo $layout->render();
+} catch (\atk4\core\Exception $e) {
+    var_dump($e->getMessage());
 
-}catch(\atk4\core\Exception $e){ 
-    var_Dump($e->getMessage());
-
-    var_Dump($e->getParams());
+    var_dump($e->getParams());
     throw $e;
 }
-
