@@ -1,11 +1,14 @@
 <?php
+
 namespace atk4\ui;
 
-class Buttons extends View {
+class Buttons extends View
+{
     public $_class = 'buttons';
 
-    function renderView() {
-        if ($this->content) { 
+    public function renderView()
+    {
+        if ($this->content) {
             $this->addClass($this->content);
             $this->content = false;
         }
