@@ -2,4 +2,6 @@
 
 namespace atk4\ui;
 
-class Exception extends \atk4\core\Exception {}
+class Exception extends \atk4\core\Exception
+{
+}
