@@ -1,8 +1,10 @@
 <?php
+
 namespace atk4\ui;
 
-class Icon extends View {
-    public $template = 'icon.html';
+class Icon extends View
+{
+    public $defaultTemplate = 'icon.html';
 
     public $content = 'circle help';
 }
