@@ -1,7 +1,8 @@
 <?php
+
 namespace atk4\ui;
 
-class Label extends View {
-    public $_class = 'label';
-
+class Label extends View
+{
+    public $ui = 'label';
 }
