@@ -1,10 +1,18 @@
 <?php
+
 namespace atk4\ui;
 
-class Text extends View {
-    public $template = false;
+class Text extends View
+{
+    public $defaultTemplate = false;
 
-    function render() {
+    public function render()
+    {
+        return $this->content;
+    }
+
+    public function getHTML()
+    {
         return $this->content;
     }
 }
